Moved closed caption fonts loading to `display.ts` and renamed the json file
