+++ src/api/index.ts
@@ -50,6 +50,7 @@ import {
     redrawDisplay,
     clearDisplay,
     statsUpdate,
+    loadCaptionsFonts,
     setDisplayState,
     setCaptionMode,
     setCaptionStyle,
@@ -74,7 +75,6 @@ import {
     handleVideoEvent,
     initVideoModule,
     isVideoMuted,
-    loadCaptionsFonts,
     muteVideo,
     resetVideo,
     subscribeVideo,
