+++ src/core/common/fonts/system-fonts.json
@@ -1,5 +1,5 @@
 {
-    "fonts": {
+    "closedCaptions": {
         "cc-serif-fixed": "Colon Mono W01 Regular.ttf",
         "cc-serif": "Colon W01 Regular.ttf",
         "cc-sans-serif-fixed": "RobotoMono-Regular.ttf",
